Refresh data; add States
Okay, so this should have been split up. Sue me.

Here we add a States view. To be fair, we probably should stop
doing this through an API and just generate it from the server,
but that's okay. We can think about that.

More usefully, we now download data every so often and update it.
Right now we don't bother writing it out to update our default
data, and we probably don't need to given the update loop. Something
to consider though.

We now have live data; we'll have to figure out how we want to
handle that.