States: Pass in a start date to the API
Eventually we'll want to do more than just provide a holistic view of
all cases for all states. It'd be nice to pick a date to look at how
the spread of viruses is affecting states - particularly as New York
dwarfed the spread early on but is generally doing better. This patch
sets that up, although I think there's probably more useful work to
do in just getting the daily stats right for the US.